requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langgraph>=0.2.0
chromadb>=0.5.0
python-dotenv>=1.0.1
pydantic>=2.7.0
tiktoken>=0.7.0